backend/requirements.txt
# FastAPI & Server
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.35
aiosqlite==0.20.0

# ML & Data
tensorflow==2.16.2
scikit-learn==1.5.2
pandas==2.2.3
numpy==1.26.4
xgboost==2.1.1
prophet==1.1.6

# Stock & AI
yfinance==0.2.43
google-generativeai==0.8.3
crewai==0.80.0
crewai-tools==0.14.0

# Visualization
matplotlib==3.9.2

# Utils
python-dotenv==1.0.1
pydantic[email]==2.9.2
